Use non-blockign random generator.
In certain containerized environments the default random number
generator can get starved of entropy and it takes a very long time to
produce random bytes (as in 10s of seconds). To avoid this we can use a
non-blocking random generator instead.

More info: https://medium.com/@TamasPolgar/hack-how-to-use-securerandom-with-kubernetes-and-docker-a375945a7b21